Split dynamic models into domain
