+++ TheBackend.Admin/Pages/Models.razor
@@ -1,6 +1,6 @@
 @page "/models"
 @inject ApiClient Api
-@using TheBackend.DynamicModels
+@using TheBackend.Domain.Models
 
 <h1 class="text-2xl font-bold mb-4">Models</h1>
