+++ TheBackend.DynamicModels/DynamicDbContextService.cs
@@ -8,6 +8,7 @@ using System.Diagnostics;
 using System.Reflection;
 using System.Runtime.Loader;
 using System.Text;
+using TheBackend.Domain.Models;
 using System.Linq;
 using System.Collections.Generic;
